requirements.txt.txt
streamlit
mysql-connector-python
pandas
python-telegram-bot